Update LICENSE
Updated the license to restrict unauthorized commercial use.

I chose a new license (CC BY-NC-SA 4.0) to prevent others from using this code for commercial purposes without my knowledge or consent. The code may still be used, adapted, and studied for non-commercial purposes, as long as it is not copied 1:1 or used in a way that generates profit. Attribution is required, and any derivative works must be shared under the same license terms.